synth: the engine thinks in bytes too
The byte reaches the far end now. A voice remembers the channel, key
and velocity it was struck with as the bytes they are; a channel's
controllers, pedals, pressures and per-note drum edits take bytes; the
modulators and the region key/velocity tests take bytes. Nothing is
carried wider than it is.

Where the maths genuinely wants a sign, it widens at the point of
computation and the reason is written down: the envelope's
cents * (60 - key) is negative for most of the keyboard, portamento's
glide offset is a signed distance in semitones, pitch bend is fourteen
bits centred on 8192, the GS tone modifies are relative about 40H, and
a shaped velocity leaves both ends of the byte before it is clamped
back inside them. The guards that mattered stay: a u8 still reaches
255, so the 7-bit tables keep their bounds checks.

The two channel accessors that had been left taking usize -- bank/patch
and the raw CC -- take the byte as well, so the whole channel-scoped
surface is one shape rather than three.

Bit-exact, and shown to be: tests/bitcheck_probe.rs drives notes across
the whole key range, every controller the dispatch reads, pitch-bend
sweeps, RPN and NRPN data entry, portamento, mono mode, a drum part
with per-note NRPN edits and both kinds of pressure, then hashes every
sample of both channels. 9cd3e866de23b7e3 before this commit and after
it. It stays as an oracle for the next refactor that must not change a
sample, ignored by default and printing rather than asserting: the
number is only comparable against itself on one machine, the envelope
maths going through the platform's own exp.

+++ src/synth/modulator.rs
@@ -186,16 +186,16 @@ impl ModSource {
 pub(crate) fn source_raw(
     source: ModSource,
     channel: &crate::synth::channel::Channel,
-    key: i32,
-    velocity: i32,
+    key: u8,
+    velocity: u8,
 ) -> f32 {
     if source.is_cc {
         return channel.get_cc(source.index) as f32 / 127.0;
     }
     match source.index {
         general_controller::NONE => 1.0,
-        general_controller::NOTE_ON_VELOCITY => velocity as f32 / 127.0,
-        general_controller::NOTE_ON_KEY => key as f32 / 127.0,
+        general_controller::NOTE_ON_VELOCITY => f32::from(velocity) / 127.0,
+        general_controller::NOTE_ON_KEY => f32::from(key) / 127.0,
         general_controller::POLY_PRESSURE => channel.get_poly_pressure(key) as f32 / 127.0,
         general_controller::CHANNEL_PRESSURE => channel.get_channel_pressure() as f32 / 127.0,
         general_controller::PITCH_WHEEL => channel.get_pitch_bend_raw(),
@@ -210,8 +210,8 @@ impl Modulator {
     pub(crate) fn contribution(
         &self,
         channel: &crate::synth::channel::Channel,
-        key: i32,
-        velocity: i32,
+        key: u8,
+        velocity: u8,
     ) -> f32 {
         let primary = ModSource::from_operator(self.source);
         let mut value = source_raw(primary, channel, key, velocity);
@@ -245,7 +245,7 @@ impl Modulator {
     /// The contribution of a modulator whose sources are all fixed at
     /// note-on. Sources that would need live channel state read as zero,
     /// so this is only meaningful for modulators [`is_static`] accepts.
-    pub(crate) fn static_contribution(&self, key: i32, velocity: i32) -> f32 {
+    pub(crate) fn static_contribution(&self, key: u8, velocity: u8) -> f32 {
         let fixed = |op: u16| {
             let m = ModSource::from_operator(op);
             let raw = if m.is_cc {
@@ -253,8 +253,8 @@ impl Modulator {
             } else {
                 match m.index {
                     general_controller::NONE => 1.0,
-                    general_controller::NOTE_ON_VELOCITY => velocity as f32 / 127.0,
-                    general_controller::NOTE_ON_KEY => key as f32 / 127.0,
+                    general_controller::NOTE_ON_VELOCITY => f32::from(velocity) / 127.0,
+                    general_controller::NOTE_ON_KEY => f32::from(key) / 127.0,
                     _ => 0.0,
                 }
             };
